refactor: extract motion command into ai-motion.ts (-188 lines from ai.ts)
Establishes the module extraction pattern for splitting ai.ts:
each sub-module exports its execute functions/types AND a
registerXxxCommand(parentCmd) function that ai.ts calls.

ai-motion.ts (~213 lines):
  - MotionCommandOptions, MotionCommandResult interfaces
  - executeMotion() function (used by agent tools)
  - registerMotionCommand(cmd) — registers 'vibe ai motion'

ai.ts:
  - Imports and re-exports executeMotion/types for backward compat
  - Calls registerMotionCommand(aiCommand) in place of the old block
  - Shrinks from 12,938 → 12,760 lines (-178 net)

All 283 tests pass.
